Segregate nightly from stable.
