+++ Cargo.toml
@@ -15,5 +15,5 @@ edition = "2018"
 travis-ci = { repository = "phaazon/try-guard", branch = "master" }
 
 [features]
-default = ["try-trait"]
-try-trait = []
+default = []
+test-nightly = []
